Move buildRegex back into matchForApp
